Refactor bzl_lock_generator.py, splitting it into separate resolve and render stages.
There are now two stages:
* raw_lock_resolver.py takes a translated lock (now called a RawLockSet) and combines
  it with environment definitions and other parameters to output a ResolvedLockSet.
* resolved_lock_renderer.p takes a ResolvedLockSet and outputs a .bzl file, like previously.
* bzl_lock_generator.py still exists as a small wrapper around the two new stages.

This change is necessary for upcoming bzlmod work.
